# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Sistema de Inventario",
        "description": "Documentación oficial de la API para la gestión de inventarios, áreas y bienes.",
        "contact": {
            "email": "soporte@tuempresa.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8081/api",
            "description": "Servidor Principal de la API"
        }
    ],
    "paths": {
...
        "/catalogo-cucop": {
            "get": {
                "tags": [
                    "Catálogo CUCOP"
                ],
                "summary": "Listar y buscar en el catálogo",
                "description": "Listar Catálogo (Búsqueda Exacta)\n\nObtiene una lista paginada de registros. Permite buscar por coincidencia exacta en 'camb' o 'clave_cucop'.",
                "operationId": "044dc182957a63eaa71ac08348ea86c8",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Término de búsqueda (CAMB o Clave CUCOP exacta)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de registros paginada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "total": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Catálogo CUCOP"
                ],
                "summary": "Crear un nuevo registro CUCOP/CAMB",
                "description": "Crear Registro en Catálogo",
                "operationId": "20b12fe126225b8d45dcbb8869f034b7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "tipo",
                                    "clave_cucop",
                                    "descripcion"
                                ],
                                "properties": {
                                    "tipo": {
                                        "description": "Requerido por validación",
                                        "type": "string",
                                        "example": "1"
                                    },
                                    "clave_cucop": {
                                        "type": "integer",
                                        "example": 12345678
                                    },
                                    "partida_especifica": {
                                        "type": "string",
                                        "example": "21101"
                                    },
                                    "clave_cucop_plus": {
                                        "type": "string"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Material de oficina"
                                    },
                                    "nivel": {
                                        "type": "string",
                                        "example": "5"
                                    },
                                    "camb": {
                                        "type": "string",
                                        "example": "C12345"
                                    },
                                    "unidad_medida": {
                                        "type": "string",
                                        "example": "Pieza"
                                    },
                                    "tipo_contratacion": {
                                        "type": "string",
                                        "example": "Adquisiciones"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Registro creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                }
            }
        },
        "/catalogo-cucop/{id}": {
            "get": {
                "tags": [
                    "Catálogo CUCOP"
                ],
                "summary": "Obtener detalles de un registro",
                "description": "Ver Registro",
                "operationId": "ffdf707c391a93a2fca1237bb98217e0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Datos del registro"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Catálogo CUCOP"
                ],
                "summary": "Actualizar un registro existente",
                "description": "Actualizar Registro",
                "operationId": "36da6f1d22cd82fc55830d2004d2eb8b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "tipo": {
                                        "type": "string"
                                    },
                                    "clave_cucop": {
                                        "type": "integer"
                                    },
                                    "partida_especifica": {
                                        "type": "string"
                                    },
                                    "descripcion": {
                                        "type": "string"
                                    },
                                    "camb": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Registro actualizado"
                    },
                    "422": {
                        "description": "Error de validación o clave duplicada"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Catálogo CUCOP"
                ],
                "summary": "Eliminar un registro",
                "description": "Eliminar Registro",
                "operationId": "b117349011a44cb2bc01f040d77e2242",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Eliminado exitosamente"
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                }
            }
        },
...
        "/formularios/departamentos": {
            "get": {
                "tags": [
                    "Formularios"
                ],
                "summary": "Obtener opciones para crear/editar Departamento",
                "description": "Obtener opciones para el formulario de Departamentos\n\nDevuelve los datos necesarios para poblar los selectores:\nlista de Áreas disponibles y lista de Jefes de Departamento (responsables).",
                "operationId": "getDepartamentoFormOptions",
                "responses": {
                    "200": {
                        "description": "Operación exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "areas": {
                                            "description": "Lista de todas las áreas",
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "area_nombre": {
                                                        "type": "string",
                                                        "example": "Subdirección Administrativa"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "responsables": {
                                            "description": "Lista de resguardantes con rol de Jefe de Departamento",
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 5
                                                    },
                                                    "res_nombre": {
                                                        "type": "string",
                                                        "example": "María González"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                }
            }
        },
...
    },
...
}